Added: ChatBox component then transferred all data from the ChatDashboard and received data as arguments
